UI updates
Modified Textures so that textures are set after the window is setup.
(Since the class is all static references no code using it [Assuming in
rendering] should cause it to crash since it isn't run before the window
is set up anyways)
Added container component, started work on the in-game HUD.
Commiting for the few important changes that will cause code-base
changes for other devs.