Changing the changes & removing redundancies
Removed commented out code in Server because of redundancy. It's in the
UDPClient class now.
Removed commented out code in UnitConstants. It can be found in older
comits and the data has already been ported to EnumFaction.
Removed ClientServer because it is outdated and will be remade in the
future if necessary.
Commented out inappropriate usage of Game in AIMovementController
Moved EnumFaction to com.github.brigade.unit.data