Line Of Sight (not much)
Added getSizeLOS method into LineOfSight which simply returns an int
representing the number of blocks in a unit's line of sight (useful for
deciding on array sizes)

Also completely removed constructor from AIMovementController as there
is currently no need for it and added in the LineOfSight class. Began
work on unit los analysis as well.