Cleaned up AI code
AIMovementController:
removed unnecessary code, commented some, and fixed loyalty calculation.
The last thing to be done would be to make an actual move method (that
change the unit's position) and throw it in the place of the TODO
comment.

UnitData:
Changed the loyalty index from an int to a double.