Conquest and Organization
I added a new folder in the res folder called Tiles, to help organize
and clean up some of the textures we have in there. I obviously threw
all of the tile textures into the folder and I replaced the grass
texture with the one that was created (and uploaded to google docs)

GroupOverflowException:
This exception is to be used/called when there is a possibility of too
many units being assigned to a group.

CriticalPoint:
This is basically a MapPoint object but on the Conquest map. It contains
the occupying army, and the EnumFaction that controls it.

Textures:
I simply changed all of the paths of the Tile Textures so that they
would show up on the map (as opposed to throwing errors)

UnitCommander:
removed a finished TODO

UnitGroup:
This is an abstract class that will be used for all forms of "groups" of
soldiers from small squads to large battalions. A UnitGroup can hold
other UnitGroups so long as the sum of the units in the child UnitGroups
does not exceed the limit of the parent UnitGroup. The UnitGroup can
also hold individual units. To keep things clean, when a child UnitGroup
is added, the units in the child are not added to the parent's list of
units. Instead, to access the child's units you must access the child.